utils: remove ability to raise capabilities in async context
Raising the thread capabilities before an .await point is inherently
unsafe, as .await can suspend the current thread with elevated
privileges, and the privilege dropping can occur later on a different
thread.

Given that, this commit removes the `raise_async()` function, which
is fundamentally unfixable. Going forward, when elevated privileges
are required for an async operation, use runtime primitives to execute
the async code in blocking mode, combined with the standard `raise()`
function.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>