protocols: move configuration structs to separate file
In the long-term we want to automate the generation of configuration
structs from the YANG modules. As a preliminary step towards that
direction, move all protocol configuration structs to the appropriate
northbound/configuration.rs file. This also has the additional benefit
of separating boilerplate code, including defaults initialization,
from the core protocol implementations.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>