ldp: remove typestate for protocol instance
Initially, the typestate pattern was implemented for protocol instances
with the intention of simplifying the codebase. It turns out, however,
that it had the opposite effect, introducing unnecessary complexity.

The typestate pattern is better suited for simpler types due to its
limited support within the Rust language.

This refactor eliminates the `Instance` typestate and introduces the
use of a view struct (`InstanceUpView`), consistent with the other
protocol implementations.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>