routing, bgp: implement recursive nexthop resolution
This commit introduces an initial implementation of recursive nexthop
resolution.

Recursive nexthops are resolved upon the addition or update of their
corresponding routes in the RIB. Currently, only one level of recursion
is supported. Routing protocols installing routes with recursive
nexthops should implement next-hop tracking to reinstall/uninstall
the affected routes whenever the resolving route for the recursive
nexthop changes.

When displaying the YANG-modeled RIB operational data, only the
recursive next-hop is currently shown, and the resolved next-hops are
not.

Note: Some code duplication exists for IPv4 and IPv6 in the netlink.rs
file. This duplication is expected to be resolved in the near future
following an update to the rtnetlink crate.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>