routing: ensure disabled protocols are not compiled
Previously, building holod with --no-default-features resulted in
the same binary size as when all features were enabled. The issue
was due to holo-routing not properly handling disabled features,
unlike holo-daemon. This commit fixes that problem.

Now, in the release profile, the binary size of holod on my machine
is 14MB with --no-default-features, compared to 22MB for a full build.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>