yang: prepare crate for publishing
Add the description field as it's required by crates.io to publish
crates.

The `holo-yang` crate needs to be published since it's used by
`holo-cli`, and `holo-cli` is going to be moved to a separate
repository (see #12).

Eventually, we want all Holo crates on crates.io, but it's too early
for that now.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>