daemon: make tokio-console integration an optional feature
The tokio-console integration is excellent for debugging but shouldn't
be needed by regular users. Using cargo-tree, it's possible to see that
the console-subscriber crate pulls in several additional dependencies,
which impact the final binary size and build time as well. As such,
make tokio-console integration optional so that this cost is incurred
only when necessary.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>