system: introduce the `holo-system` component
Introduce `holo-system`, implementing the following subset of the
ietf-system YANG module:

module: ietf-system
  +--rw system
  |  +--rw contact?    string
  |  +--rw hostname?   inet:domain-name
  |  +--rw location?   string
  +--ro system-state
     +--ro platform
     |  +--ro os-name?      string
     |  +--ro os-release?   string
     |  +--ro os-version?   string
     |  +--ro machine?      string
     +--ro clock
        +--ro current-datetime?   yang:date-and-time
        +--ro boot-datetime?      yang:date-and-time

The main motivation for this component is to provide a centralized
place for storing hostname information and sharing it with protocol
instances (both IS-IS and OSPF have TLVs for flooding hostname
information). Additionally, including the hostname in the configuration
is necessary for implementing a persistent hostname in holo-cli.

Closes #20

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>