isis: embed interface index in the struct
Embed the interface index directly in the interface struct. Since
the interface index in the arena is immutable after the interface
is created, this change simplifies the API. Many APIs can now return
a simple interface reference instead of a tuple containing both the
interface index and the reference.

This pattern can be applied to other collections as well, like
adjacencies and LSP entries, but one step at a time.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>