isis: implement cryptographic authentication
This commit adds support for both clear-text authentication, as
specified by the ISO standard, and cryptographic authentication,
as specified by RFC 5304.

This includes the ability to authenticate all PDU types: Hello,
Sequence Number, and Link State PDUs. Both manual keys and keychains
are supported.

Support for RFC 5310 ("IS-IS Generic Cryptographic Authentication")
will be left for a second moment.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>