*: replace ibus broadcast channel with mpsc channels
This is the beginning of a major overhaul in the ibus infrastructure.

Using a broadcast channel for ibus communication has the benefit of
being simple: every message sent is received by all core components
(e.g., holo-routing, holo-interface, etc.) and protocol instances.

However, the downside is that broadcast messages are suboptimal,
as many receivers are not interested in some messages and will
simply drop them. Additionally, the broadcast channel is unreliable
and may start dropping messages once it is full. This is often the
case when testing BGP, where hundreds of thousands of routes can
be installed in a short timeframe.

In this commit, we transition to a more efficient network of mpsc
channels. Each base component and protocol instance now has a
dedicated transmission endpoint to all other base components. This
allows unicast communication, such as holo-bgp directly installing
routes to holo-routing. For certain communication patterns,
like interface system updates, holo-interface sends messages to
holo-routing, which then relays them to the protocol instances.

This change lays the foundation for future improvements, such as
implementing a pub-sub pattern, where components can request and
publish data selectively for more efficient communication.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>