isis: run SPF if adjacency addresses change
Adjacency addresses are used to determine route next-hops, so when
they change, we need to trigger SPF to update the routing table.

Usually, changes in addresses will also update the neighbor LSP,
but not always. So, we manually trigger SPF to make sure our routes
stay valid when the adjacency addresses change.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>