isis, vrrp, utils: add safe `LinkAddr` constructor
The nix crate's `LinkAddr` doesn't have a proper constructor, just
`LinkAddr::from_raw()`, which requires a raw pointer and is unsafe.

Add `LinkAddrExt::new()` in holo-utils to provide a safe way to
create `LinkAddr` instances. This removes the need for unsafe code
in protocol crates.

With this, holo-isis and holo-vrrp can inherit lint settings from
the workspace, which includes forbidding unsafe code.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>