isis: prevent installation of locally attached networks
This commit reverts e1b620e, as merely ignoring locally attached
networks in the local LSP is ineffective. Neighboring routers also
advertise these networks, meaning they are still considered for
installation. The only exception is loopback networks.

To address this, we need to track and mark all connected routes,
ensuring they are not installed in the global RIB.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>