isis: always set L1-IS bit in L2 LSPs
While re-reading the excellent "The Complete IS-IS Routing Protocol"
by Hannes Gredler, I noticed that the book highlights an oddity in the
ISO specification that we previously missed: the "Level 1 Intermediate
System" LSP bit must always be set, even in L2-only systems.

It doesn't make much sense, but it is what it is. Fix the code to
always set this bit and regenerate test data accordingly.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>