isis: add support for SR Adjacency-SIDs
Install IPv4 and IPv6 Adj-SIDs for all established adjacencies.

Advanced features, such as configurable Adj-SID weights and groups,
are left to be implemented later.

As a side note, the ietf-isis-sr-mpls YANG module was once again
slightly modified due to an incorrect augmentation. This issue will
be reported and should be addressed in the next draft version.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>