isis: check LSP lifetime when detecting content changes
When a new LSP is installed into the LSDB, we compare it with the
previous instance to determine if it represents a content change
(and possibly a topology change).

This commit updates the logic to also account for the LSP lifetime:
if the new LSP is expired and the previous one was not (or vice versa),
we treat it as a topology change and trigger a full SPF run.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>