rip: refactor code by removing typestate structs
In the early days of the project, we used the typestate pattern
to provide additional compile-time checks, such as requiring an
`InterfaceUp` parameter to guarantee at compile time that the interface
was operational. However, it turns out that the cons of this approach
outweigh the pros, as the added complexity is too high.

In this refactor, we remove the typestate pattern from the `Instance`
and `Interface` structs and adapt the rest of the code accordingly. We
also introduce a view struct (`InstanceUpView`), consistent with
other protocol implementations.

Unfortunately, the typestate pattern is better suited to simpler
types due to its limited support within the Rust language.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>