routing: fix uninstallation of blackhole routes via netlink
The netlink code was failing to uninstall blackhole routes because
the `RouteMessageBuilder::new()` API sets the route type to
`RouteType::Unicast` by default, which excludes routes of other
types, such as blackhole routes.

To fix this, the route type is now set to `RouteType::Unspec`
so the RTM_DELROUTE message works regardless of the route type.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>