isis: implement the multi-topology TLVs
Implement support for the following TLVs as defined in RFC 5120:
- TLV 229: Multi-Topology (MT) Identifiers
- TLV 222: MT Intermediate System Neighbors (MT-ISN)
- TLV 235: MT IP Reachability
- TLV 237: MT IPv6 Reachability

The latter three TLVs share the same structure as their single-topology
counterparts, with the addition of a 2-byte MT-ID. Existing TLV
structs are reused to avoid code duplication.

Also add unit tests and northbound state callbacks for MT TLVs
installed in the LSDB.

Support for configuring and using multi-topology routing will be
added in a subsequent commit.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>