yang, isis: use identities to configure multi-topology
In the IETF IS-IS YANG module, each configured IS-IS topology must
point to a global RIB table. This is because, in theory, one can
configure multiple topologies for the same address family, and
the computed routes for those topologies must be kept separate. In
practice, however, most vendors only support configuring the IPv6
unicast topology, which is also the case for us.

For the sake of simplicity, we had previously deviated from the IS-IS
YANG module by using a string to represent the topology name, instead
of an identityref to a RIB name. In this commit, we change the string
to an identityref to an MT topology name. This makes it easier for
users to identify valid values and also allows CLI applications to
support auto-completion of MT topology names.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>