isis: update BFD state handling to accommodate multi-topology
When BFD is enabled on an interface, separate IPv4 and IPv6 BFD
sessions are registered for each dual-stack adjacency. Previously,
the adjacency was torn down entirely when any one of those sessions
went down.

According to RFC 5882, Section 10.1.2 ("Reaction to BFD State
Changes"), IS-IS implementations must take special care when operating
in multi-topology mode. In such cases, if a BFD session fails, we
should not tear down the entire adjacency. Instead, we now remove
IS-reachability information only from the affected topologies in the
local LSP. For example, if the IPv4 BFD session goes down, the standard
topology is affected, but the IPv6 unicast topology remains intact.

The adjacency is now only fully torn down when all associated BFD
sessions (IPv4 and/or IPv6) are down.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>