isis: add configuration option to control N-flag on loopback routes
Previously, the N-flag was always set on host routes (/32 or /128)
associated with loopback interfaces. This change introduces a node-flag
interface configuration option (default: true) that allows this
behavior to be configured.

This is useful in scenarios where the N-flag should be unset, such
as on Anycast-SIDs (otherwise TI-LFA won't work correctly).

The node-flag option is defined in the IETF IS-IS YANG module and is
gated by the node-flag feature, which is now implemented.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>