isis: implement protection against corrupted LSP lifetimes
Implement protection against corrupted Remaining Lifetime values
in LSPs as described in RFC 7987. When a received LSP is accepted
(i.e. newer or not present in the LSDB), and it is not expired,
its Remaining Lifetime is reset to the configured maximum before
installation.

As suggested by the RFC, the original Remaining Lifetime value received
in the LSP is retained and exposed as YANG-modeled operational data.
This serves as a diagnostic aid when troubleshooting LSP lifetime
corruption issues.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>