protocols: remove async-trait dependency
async fn in traits was stabilized some time ago, so we don't need the
async-trait crate anymore.

Ideally, we could also drop the `MessageReceiver` trait and
have `ProtocolInputChannelsRx` implement `futures::Stream<Item =
ProtocolInputMsg>`. The problem is that we currently rely on `select!`
with the `biased` option to keep test results deterministic. Long
term it would be nice to find a way to get deterministic tests while
using fair polling of all protocol events.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>