isis: resolve level type for newly configured interfaces
Fix a case where the interface level type was not correctly resolved
when the interface was added in a separate transaction from the IS-IS
instance and no explicit level type was configured. Without this
fix, the interface could operate as L1/L2 even if IS-IS instance was
configured as single-level.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>