isis: resign from DIS role when there are no adjacencies left
Fix a corner case where we are the DIS and all adjacencies go down
(timeouts, clear RPC, etc). In this case, we need to resign from
the DIS role, which means stop sending CSNP PDUs and purge the
pseudonode LSP.

Also add a conformance test for this scenario.

Issue discovered with IxANVL.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>