isis: add support for MSD signaling
Add support for MSD (Maximum SID Depth) signaling as specified in
RFC 8491.

The ietf-isis-msd YANG module (currently in draft) was deviated
to use integer types instead of identityrefs for MSD types. This
allows unknown MSD types to be displayed, consistent with the OSPF
module. This will be reported for consideration in the next version
of the draft.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>