isis: skip overload bit check for the local LSP
During SPF, the overload bit check must be skipped for the local
LSP. If the bit is set and the check is applied, the SPT ends
up empty.

In the ISO specification, TENT is preloaded from the adjacency
database.  In this implementation, local adjacency information is
fetched directly from the local LSP instead so the algorithm can
be run with any node as the root.  As such, this special case for
the overload bit check is not documented in the ISO standard but
is necessary here.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>