isis: implement PSNP fallback for flooding failures
This commit implements the missing piece of
draft-ietf-lsr-distoptflood-12, which is the handling of flooding
failures specified in section 1.2.4.

The strategy is to send, after a short period of time, a PSNP
containing all LSPs that have not been reflooded over an interface,
to ensure that the neighbor(s) request the missing LSPs in the event
of a flooding failure. While CSNPs already serve as a fallback
for flooding failures, this strategy is intended to speed up
convergence times.

This implementation reuses partialSNPInterval and SSNflags from the
base protocol to trigger these PSNPs. Testing confirms that this
works well. In rare cases, if the PSNP interval fires immediately
after an LSP is pruned from flooding on the interface, an unnecessary
reflood may occur. This behavior is harmless but suboptimal. Using
a dedicated timer for PSNP-based flooding failure recovery may be
considered in the future.

Signed-off-by: Renato Westphal <renato@opensourcerouting.org>